requirements.txt
alembic==1.11.1
antiorm==1.2.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
autopep8==2.0.2
azure-core==1.28.0
azure-storage-blob==12.17.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
config==0.5.1
cryptography==41.0.1
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
ecdsa==0.18.0
executing==1.2.0
fastapi==0.99.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.3.2
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==3.0.1
mysql-connector-python==8.0.33
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
PyJWT==2.7.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-utils==3.7.0
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
repoze.lru==0.7
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Routes==2.5.1
rsa==4.9
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
virtualenv @ git+https://github.com/pypa/virtualenv.git@9f9dc6250fc88e92b1ca6206429966788846d696
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.4
yarg==0.1.9
zipp==3.15.0